server/src/requirements.txt
python-dotenv==1.0.1
fastapi==0.114.0
openai==1.44.1